fix: eslint errors
