+++ src/pages/feeds.tsx
@@ -1,23 +1,23 @@
 import {
-    Flex,
-    Box,
-    Button,
-    useDisclosure,
-    Modal,
-    ModalContent,
-    ModalHeader,
-    ModalFooter,
-    Text,
-    HStack,
-    Popover,
-    PopoverTrigger,
-    PopoverContent,
-    PopoverHeader,
-    PopoverFooter,
-    Tag,
-    Icon,
-    BoxProps,
-    useToast
+  Flex,
+  Box,
+  Button,
+  useDisclosure,
+  Modal,
+  ModalContent,
+  ModalHeader,
+  ModalFooter,
+  Text,
+  HStack,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverHeader,
+  PopoverFooter,
+  Tag,
+  Icon,
+  type BoxProps,
+  useToast
 } from '@chakra-ui/react';
 import Layout from '~/layout/index';
 import { useForm } from 'react-hook-form';
@@ -26,178 +26,224 @@ import { Header } from '~/components/Header';
 import { SlOptions, SlTrash } from 'react-icons/sl';
 import { deleteFile } from '~/utils/file';
 import { api } from '~/utils/api';
-import AddFeed from './add-feed';
-import EditFeed from './edit-feed';
+import AddFeed from '~/components/feeds/add-feed';
+import EditFeed from '~/components/feeds/edit-feed';
+import { withSession } from '~/server/auth/withSession';
+import AdminRoute from '~/layout/AdminRoute';
+import { useSession } from 'next-auth/react';
+
+export const getServerSideProps = withSession({ force: true });
 
 export interface FeedProps {
-    id: number;
-    content: string;
-    url?: string | undefined;
-    filePath?: any;
+  id: number;
+  content: string;
+  url?: string | undefined;
+  filePath?: any;
 }
 
+// eslint-disable-next-line react/display-name
 const PopoverBox = forwardRef<HTMLDivElement, BoxProps>((props, ref) => {
-    return (
-        <Box ref={ref} {...props}>
-            {props.children}
-        </Box>
-    );
+  return (
+    <Box ref={ref} {...props}>
+      {props.children}
+    </Box>
+  );
 });
 
 export default function Feeds() {
-    const toast = useToast();
-    const { reset, setValue, getValues } = useForm<FeedProps>({
-        mode: 'onSubmit',
-        defaultValues: {
-            id: undefined,
-            content: '',
-            url: '',
-        }
-    });
-
-    const { isOpen: isRemoveOpen, onOpen: onRemoveOpen, onClose: onRemoveClose } = useDisclosure();
-    const getFeedsList = api.feeds.adminGetFeeds.useQuery();
-    const deleteFeed = api.feeds.adminDeleteFeed.useMutation()
-    const [fetchCount, setFetchCount] = useState<number>(0);
-    const feedsList = getFeedsList.data
-    const [shouldRefetch, setShouldRefetch] = useState(false);
-    
-    const increment = () => {
-        setFetchCount(fetchCount + 1)
+  const { data: session } = useSession();
+  const toast = useToast();
+  const { reset, setValue, getValues } = useForm<FeedProps>({
+    mode: 'onSubmit',
+    defaultValues: {
+      id: undefined,
+      content: '',
+      url: ''
     }
-    console.log('jumlah', fetchCount)
-    useEffect(() => {
-        if (shouldRefetch) {
-            const delay = 500;
-            const timeoutId = setTimeout(() => {
-                getFeedsList.refetch();
-                setShouldRefetch(false);
-            }, delay);
-            
-            return () => clearTimeout(timeoutId);
-        }
-    }, [shouldRefetch, getFeedsList]);
-    
-    useEffect(() => {
-        setShouldRefetch(true);
-    }, [fetchCount]);
-    
-    const handleDelete = (idValue: number) => {
-        setValue('id', idValue)
+  });
+
+  const {
+    isOpen: isRemoveOpen,
+    onOpen: onRemoveOpen,
+    onClose: onRemoveClose
+  } = useDisclosure();
+  const getFeedsList = api.feeds.adminGetFeeds.useQuery();
+  const deleteFeed = api.feeds.adminDeleteFeed.useMutation();
+  const [fetchCount, setFetchCount] = useState<number>(0);
+  const feedsList = getFeedsList.data;
+  const [shouldRefetch, setShouldRefetch] = useState(false);
+
+  const increment = () => {
+    setFetchCount(fetchCount + 1);
+  };
+  useEffect(() => {
+    if (shouldRefetch) {
+      const delay = 500;
+      const timeoutId = setTimeout(() => {
+        void getFeedsList.refetch();
+        setShouldRefetch(false);
+      }, delay);
+
+      return () => clearTimeout(timeoutId);
     }
+  }, [shouldRefetch, getFeedsList]);
 
-    const handleSubmitDelete = async (itemId: number, url: string | null) => {
-        try {
-            deleteFeed.mutate({ feedId: itemId });
-            if (url) {
-                await deleteFile(url);
-            }
-        } catch (error) {
-            console.log(error);
-        }
-        setFetchCount(fetchCount + 1);
-        reset();
-        onRemoveClose();
-    };
+  useEffect(() => {
+    setShouldRefetch(true);
+  }, [fetchCount]);
 
-    const getTimeLabel = (timestamp: Date): string => {
-        const currentTime = new Date();
-        const uploadTime = new Date(timestamp);
-        const timeDifference = (currentTime.getTime() - uploadTime.getTime()) / (1000 * 60);
+  const handleDelete = (idValue: number) => {
+    setValue('id', idValue);
+  };
 
-        if (timeDifference < 60) {
-            return `Diunggah ${Math.floor(timeDifference)} menit yang lalu`;
-        } else if (timeDifference < 1440) {
-            return `Diunggah ${Math.floor(timeDifference / 60)} jam yang lalu`;
-        } else {
-            return `Diunggah ${Math.floor(timeDifference / 1440)} hari yang lalu`;
-        }
-    };
+  const handleSubmitDelete = async (itemId: number, url: string | null) => {
+    try {
+      deleteFeed.mutate({ feedId: itemId });
+      if (url) {
+        await deleteFile(url);
+      }
+    } catch (error) {
+      console.log(error);
+    }
+    setFetchCount(fetchCount + 1);
+    reset();
+    onRemoveClose();
+  };
 
-    return (
-        <Layout type='admin' title='Feeds' fullBg={true}>
-            <Box width='100%'>
-                <Header title='Feeds' />
-                <Box
-                    mt='1.5rem'
-                    bgColor='#ffffff'
-                    border='1px solid #CBD2E0'
-                    borderRadius='8px'
-                >
-                    <AddFeed feedChange={increment}/>
-                    {/* ========== content List ========== */}
-                    <Box overflowY='auto' width='100%' maxHeight='600px'
-                        css={{
-                            '&::-webkit-scrollbar': {
-                                width: '4px',
-                            },
-                            '&::-webkit-scrollbar-track': {
-                                width: '6px',
-                            },
-                            '&::-webkit-scrollbar-thumb': {
-                                background: '#2D3648',
-                                borderRadius: '24px',
-                            },
-                        }}
-                    >
-                        {feedsList?.slice()
-                            .sort((b, a) => new Date(a.createdAt).getTime() - new Date(b.createdAt).getTime()).map((item, index: number) => (
-                                <Box key={index} px='1.5rem' borderTop='1px solid #CBD2E0'>
-                                    <Box float='right' my='1rem' cursor='pointer'>
-                                        <Popover placement='bottom-end' autoFocus={false}>
-                                            <PopoverTrigger>
-                                                <PopoverBox>
-                                                    <Icon as={SlOptions} boxSize={6} fill='gray.300' />
-                                                </PopoverBox>
-                                            </PopoverTrigger>
-                                            <PopoverContent width='fit-content' borderColor='#000' zIndex='100000'>
-                                                <PopoverHeader borderColor='#000'>
-                                                    <EditFeed id={item.id} content={item.text} url={item.attachmentUrl!} feedChange={increment}/>
-                                                </PopoverHeader>
-                                                <PopoverFooter >
-                                                    <HStack onClick={() => {
-                                                        handleDelete(item.id);
-                                                        onRemoveOpen();
-                                                    }} >
-                                                        <Icon as={SlTrash} boxSize={6} />
-                                                        <Text>Remove</Text>
-                                                    </HStack>
-                                                    <Modal isOpen={isRemoveOpen} onClose={onRemoveClose}>
-                                                        <ModalContent>
-                                                            <ModalHeader>Are you sure to delete this feed?</ModalHeader>
-                                                            <ModalFooter>
-                                                                <Button variant='outlineBlue' mr={3} onClick={onRemoveClose}>
-                                                                    Cancel
-                                                                </Button>
-                                                                <Button variant='solidBlue' onClick={() => handleSubmitDelete(getValues('id'), item.attachmentUrl)}>
-                                                                    Confirm
-                                                                </Button>
-                                                            </ModalFooter>
-                                                        </ModalContent>
-                                                    </Modal>
-                                                </PopoverFooter>
-                                            </PopoverContent>
-                                        </Popover>
-                                    </Box>
-                                    <Box py='1rem'>
-                                        <Flex justifyContent='space-between' alignItems='center'>
-                                            <Text fontSize='2xl' fontWeight='700'>OSKM ITB 2023</Text>
-                                            <Text mr='1rem'>{getTimeLabel(item.createdAt)}</Text>
-                                        </Flex>
-                                        <Text>
-                                            {item.text}
-                                        </Text>
-                                        {
-                                            item.attachmentUrl && (
-                                                <Tag>{item.attachmentUrl}</Tag>
-                                            )
-                                        }
-                                    </Box>
-                                </Box>
-                            ))}
+  const getTimeLabel = (timestamp: Date): string => {
+    const currentTime = new Date();
+    const uploadTime = new Date(timestamp);
+    const timeDifference =
+      (currentTime.getTime() - uploadTime.getTime()) / (1000 * 60);
+
+    if (timeDifference < 60) {
+      return `Diunggah ${Math.floor(timeDifference)} menit yang lalu`;
+    } else if (timeDifference < 1440) {
+      return `Diunggah ${Math.floor(timeDifference / 60)} jam yang lalu`;
+    } else {
+      return `Diunggah ${Math.floor(timeDifference / 1440)} hari yang lalu`;
+    }
+  };
+
+  return (
+    <AdminRoute session={session}>
+      <Layout type='admin' title='Feeds' fullBg={true}>
+        <Box width='100%'>
+          <Header title='Feeds' />
+          <Box
+            mt='1.5rem'
+            bgColor='#ffffff'
+            border='1px solid #CBD2E0'
+            borderRadius='8px'
+          >
+            <AddFeed feedChange={increment} />
+            {/* ========== content List ========== */}
+            <Box
+              overflowY='auto'
+              width='100%'
+              maxHeight='600px'
+              css={{
+                '&::-webkit-scrollbar': {
+                  width: '4px'
+                },
+                '&::-webkit-scrollbar-track': {
+                  width: '6px'
+                },
+                '&::-webkit-scrollbar-thumb': {
+                  background: '#2D3648',
+                  borderRadius: '24px'
+                }
+              }}
+            >
+              {feedsList
+                ?.slice()
+                .sort(
+                  (b, a) =>
+                    new Date(a.createdAt).getTime() -
+                    new Date(b.createdAt).getTime()
+                )
+                .map((item, index: number) => (
+                  <Box key={index} px='1.5rem' borderTop='1px solid #CBD2E0'>
+                    <Box float='right' my='1rem' cursor='pointer'>
+                      <Popover placement='bottom-end' autoFocus={false}>
+                        <PopoverTrigger>
+                          <PopoverBox>
+                            <Icon as={SlOptions} boxSize={6} fill='gray.300' />
+                          </PopoverBox>
+                        </PopoverTrigger>
+                        <PopoverContent
+                          width='fit-content'
+                          borderColor='#000'
+                          zIndex='100000'
+                        >
+                          <PopoverHeader borderColor='#000'>
+                            <EditFeed
+                              id={item.id}
+                              content={item.text}
+                              url={item.attachmentUrl!}
+                              feedChange={increment}
+                            />
+                          </PopoverHeader>
+                          <PopoverFooter>
+                            <HStack
+                              onClick={() => {
+                                handleDelete(item.id);
+                                onRemoveOpen();
+                              }}
+                            >
+                              <Icon as={SlTrash} boxSize={6} />
+                              <Text>Remove</Text>
+                            </HStack>
+                            <Modal
+                              isOpen={isRemoveOpen}
+                              onClose={onRemoveClose}
+                            >
+                              <ModalContent>
+                                <ModalHeader>
+                                  Are you sure to delete this feed?
+                                </ModalHeader>
+                                <ModalFooter>
+                                  <Button
+                                    variant='outlineBlue'
+                                    mr={3}
+                                    onClick={onRemoveClose}
+                                  >
+                                    Cancel
+                                  </Button>
+                                  <Button
+                                    variant='solidBlue'
+                                    onClick={() =>
+                                      void handleSubmitDelete(
+                                        getValues('id'),
+                                        item.attachmentUrl
+                                      )
+                                    }
+                                  >
+                                    Confirm
+                                  </Button>
+                                </ModalFooter>
+                              </ModalContent>
+                            </Modal>
+                          </PopoverFooter>
+                        </PopoverContent>
+                      </Popover>
                     </Box>
-                </Box>
+                    <Box py='1rem'>
+                      <Flex justifyContent='space-between' alignItems='center'>
+                        <Text fontSize='2xl' fontWeight='700'>
+                          OSKM ITB 2023
+                        </Text>
+                        <Text mr='1rem'>{getTimeLabel(item.createdAt)}</Text>
+                      </Flex>
+                      <Text>{item.text}</Text>
+                      {item.attachmentUrl && <Tag>{item.attachmentUrl}</Tag>}
+                    </Box>
+                  </Box>
+                ))}
             </Box>
-        </Layout>
-    );
+          </Box>
+        </Box>
+      </Layout>
+    </AdminRoute>
+  );
 }
